updated profile menu
