+++ src/components/molecules/NavMenu.jsx
@@ -1,8 +1,5 @@
 import React, { useContext, useState, useEffect } from 'react';
 import AuthContext from 'context/AuthContext';
-import { signOut } from 'firebase/auth';
-import { auth } from '../../../firebase';
-import { useNavigate } from 'react-router-dom';
 
 import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
@@ -34,7 +31,6 @@ const NavMenu = styled.nav`
 const NavBar = ({ links }) => {
   const { currentUser } = useContext(AuthContext);
   const [isLoading, setIsLoading] = useState(true);
-  const navigate = useNavigate();
 
   useEffect(() => {
     // Simulate checking the currentUser state
@@ -43,20 +39,6 @@ const NavBar = ({ links }) => {
     }
   }, [currentUser]);
 
-  const handleLogout = async () => {
-    try {
-      await signOut(auth);
-      navigate('/');
-    } catch (err) {
-      console.error('Logout Error:', err);
-    }
-  };
-
-  const menuItems = [
-    { label: 'View Profile', onClick: () => console.log('View Profile') },
-    { label: 'Settings', onClick: () => console.log('Settings') },
-    { label: 'Logout', onClick: handleLogout },
-  ];
   // Show nothing until loading is complete
   if (isLoading) return null;
 
@@ -67,17 +49,7 @@ const NavBar = ({ links }) => {
           {link.label}
         </Link>
       ))}
-      {currentUser ? (
-        <ProfileDropdown
-          menuItems={menuItems}
-          avatarImage={
-            'https://storage.googleapis.com/transpiled-web/images/joshua.jpg'
-          }
-          avatarName={'Joshua'}
-        />
-      ) : (
-        <Link to="/login">Login</Link>
-      )}
+      {currentUser ? <ProfileDropdown /> : <Link to="/login">Login</Link>}
     </NavMenu>
   );
 };
